+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 7.6.0
+
+* Address Autocomplete Field: Add `closestLocationsFilter` option to allow filtering closest locations before they are saved to the form data.
+
 ## 7.5.0
 
 * Hide `submitOptions` for admins as it was confusing.